# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prediction Game API",
    "description": "API for wallet-authenticated prediction gameplay, leaderboards, rounds, and predictions. Use Swagger UI to explore endpoints and test requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste a JWT like: Bearer <token>"
      }
    },
...
  },
...
  "paths": {
...
    "/api/bets/up-down": {
      "post": {
        "summary": "Submit an UP/DOWN bet (stub)",
        "tags": [
          "bets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "address",
                  "amount",
                  "side"
                ],
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "Optional; must match JWT wallet when provided"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "UP",
                      "DOWN"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bet recorded (stub)"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid JWT"
          }
        }
      }
    },
    "/api/bets/precision": {
      "post": {
        "summary": "Submit a Precision bet (stub)",
        "tags": [
          "bets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "address",
                  "amount",
                  "predictedPrice"
                ],
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "Optional; must match JWT wallet when provided"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "predictedPrice": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bet recorded (stub)"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid JWT"
          }
        }
      }
    },
    "/api/bets/claim": {
      "post": {
        "summary": "Claim pending Soroban winnings",
        "tags": [
          "bets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "description": "Optional; must match JWT wallet when provided"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Claim recorded or submitted on-chain"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Wallet address mismatch"
          },
          "409": {
            "description": "Idempotency key conflict"
          },
          "422": {
            "description": "No claimable winnings / invalid contract state"
          },
          "503": {
            "description": "Contract interaction failed"
          }
        }
      }
    },
    "/api/bets/reconciliation": {
      "get": {
        "summary": "Query bet records and their on-chain reconciliation state",
        "description": "Admin-only. Returns bet records with their reconciliation fields (status, txHash, submittedAt) plus a per-status summary, so stub bets and their eventual on-chain transactions can be audited together.\n",
        "tags": [
          "bets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "schema": {
              "type": "string"
            },
            "description": "Filter by Stellar wallet address"
          },
          {
            "in": "query",
            "name": "roundId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by round"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "STUB",
                "SUBMITTED",
                "CONFIRMED",
                "FAILED"
              ]
            },
            "description": "Filter by reconciliation status"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching bet records, newest first, with a status summary"
          },
          "400": {
            "description": "Unknown status filter"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/api/bets/{id}": {
      "get": {
        "summary": "Get a single bet record with its reconciliation state",
        "tags": [
          "bets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bet ID returned when the bet was placed"
          }
        ],
        "responses": {
          "200": {
            "description": "Bet record"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Bet not found"
          }
        }
      }
    },
...
  }
}